rename: 파일들을 패키지로 다시 분류
