+++ src/main/java/kr/codesqaud/cafe/domain/User.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe;
+package kr.codesqaud.cafe.domain;
 
 public class User {
     private static int userNumFactory = 0;
