+++ src/main/java/kr/codesqaud/cafe/service/UserService.java
@@ -1,5 +1,7 @@
-package kr.codesqaud.cafe;
+package kr.codesqaud.cafe.service;
 
+import kr.codesqaud.cafe.domain.User;
+import kr.codesqaud.cafe.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
